Implement mass_gradient precipitation fraction.
This commit introduces a new precipitation fraction calculation method,
the "mass gradient" method, to MG2. This method uses the gradient of the
total precipitation mass mixing ratio to determine the rate at which the
precipitation fraction decreases with height in clear regions.

Note that there are two additions to the namelist (though there will be
no changes to namelist settings in a default case):

1. The string "mass_gradient" is now an acceptable setting for the
   micro_mg_precip_frac_method atm namelist variable.
2. The namelist values micro_mg_mass_gradient_alpha and
   micro_mg_mass_gradient_beta can now be set, and will be used if and
   only if the "mass_gradient" method is used.

The alpha and beta values adjust the rate at which the precipitation
fraction changes in regions of increasing or decreasing precipitation
mass, respectively. They should be regarded as tuning parameters which
control the balance of process rates affecting precipitation.

[BFB]
[FCC]